refactor(gateway): extract GatewayVoiceMixin from run.py (slice 23 of #54962)
